--- src/Author.java
@@ -1,17 +1,0 @@
-public class Author {
-    private String nameAuthor;
-    private String surname;
-
-    public Author(String nameAuthor, String surname) {
-        this.nameAuthor = nameAuthor;
-        this.surname = surname;
-    }
-
-    public String getNameAuthor() {
-        return this.nameAuthor;
-    }
-
-    public String getSurname() {
-        return this.surname;
-    }
-}
